Link `ftPr_SpecialLw` (936 bytes)
