+++ configure.py
@@ -640,7 +640,8 @@ config.libs = [
             # Jigglypuff
             Object(Matching, "melee/ft/chara/ftPurin/ftPr_Init.c"),
             Object(Matching, "melee/ft/chara/ftPurin/ftPr_SpecialHi.c"),
-            Object(NonMatching, "melee/ft/chara/ftPurin/ftPr_SpecialLw.c"),
+            Object(Matching, "melee/ft/chara/ftPurin/ftPr_SpecialLw.c"),
+            Object(NonMatching, "melee/ft/chara/ftPurin/ftPr_SpecialS.c"),
             # Luigi
             Object(Matching, "melee/ft/chara/ftLuigi/ftLg_Init.c"),
             Object(Matching, "melee/ft/chara/ftLuigi/ftLg_SpecialN.c"),
